# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/v1/roles":  {
                                        "post":  {
                                                     "operationId":  "RolesController_create_v1",
                                                     "parameters":  [

                                                                    ],
                                                     "requestBody":  {
                                                                         "required":  true,
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/CreateRoleDto"
                                                                                                                             }
                                                                                                              }
                                                                                     }
                                                                     },
                                                     "responses":  {
                                                                       "201":  {
                                                                                   "description":  ""
                                                                               }
                                                                   },
                                                     "tags":  [
                                                                  "Roles"
                                                              ]
                                                 },
                                        "get":  {
                                                    "operationId":  "RolesController_findAll_v1",
                                                    "parameters":  [
                                                                       {
                                                                           "name":  "page",
                                                                           "required":  false,
                                                                           "in":  "query",
                                                                           "description":  "Page number",
                                                                           "schema":  {
                                                                                          "minimum":  1,
                                                                                          "default":  1,
                                                                                          "type":  "number"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "limit",
                                                                           "required":  false,
                                                                           "in":  "query",
                                                                           "description":  "Number of records per page",
                                                                           "schema":  {
                                                                                          "minimum":  1,
                                                                                          "maximum":  100,
                                                                                          "default":  20,
                                                                                          "type":  "number"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "isActive",
                                                                           "required":  false,
                                                                           "in":  "query",
                                                                           "description":  "Filter by active status",
                                                                           "schema":  {
                                                                                          "example":  true,
                                                                                          "type":  "boolean"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "isSystemRole",
                                                                           "required":  false,
                                                                           "in":  "query",
                                                                           "description":  "Filter by system role status",
                                                                           "schema":  {
                                                                                          "example":  false,
                                                                                          "type":  "boolean"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "search",
                                                                           "required":  false,
                                                                           "in":  "query",
                                                                           "description":  "Search term for name or key (case-insensitive)",
                                                                           "schema":  {
                                                                                          "example":  "hr",
                                                                                          "type":  "string"
                                                                                      }
                                                                       }
                                                                   ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  ""
                                                                              }
                                                                  },
                                                    "tags":  [
                                                                 "Roles"
                                                             ]
                                                }
                                    },
                  "/api/v1/roles/{id}":  {
                                             "get":  {
                                                         "operationId":  "RolesController_findOne_v1",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "required":  true,
                                                                                "in":  "path",
                                                                                "schema":  {
                                                                                               "$ref":  "#/components/schemas/ObjectId"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  ""
                                                                                   }
                                                                       },
                                                         "tags":  [
                                                                      "Roles"
                                                                  ]
                                                     },
                                             "patch":  {
                                                           "operationId":  "RolesController_update_v1",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "id",
                                                                                  "required":  true,
                                                                                  "in":  "path",
                                                                                  "schema":  {
                                                                                                 "$ref":  "#/components/schemas/ObjectId"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "requestBody":  {
                                                                               "required":  true,
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/UpdateRoleDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  ""
                                                                                     }
                                                                         },
                                                           "tags":  [
                                                                        "Roles"
                                                                    ]
                                                       }
                                         },
                  "/api/v1/roles/{id}/status":  {
                                                    "patch":  {
                                                                  "operationId":  "RolesController_updateStatus_v1",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "id",
                                                                                         "required":  true,
                                                                                         "in":  "path",
                                                                                         "schema":  {
                                                                                                        "$ref":  "#/components/schemas/ObjectId"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "requestBody":  {
                                                                                      "required":  true,
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/UpdateRoleStatusDto"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  ""
                                                                                            }
                                                                                },
                                                                  "tags":  [
                                                                               "Roles"
                                                                           ]
                                                              }
                                                },
...
              },
    "info":  {
                 "title":  "CRM + HRMS API",
                 "description":  "Production API documentation for the CRM + HRMS platform.",
                 "version":  "1.0.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "access-token":  {
                                                                    "scheme":  "bearer",
                                                                    "bearerFormat":  "JWT",
                                                                    "type":  "http",
                                                                    "description":  "Enter JWT access token"
                                                                }
                                           },
                       "schemas":  {
                                       "CreateRoleDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "name":  {
                                                                                             "type":  "string",
                                                                                             "example":  "HR Manager",
                                                                                             "description":  "The name of the role"
                                                                                         },
                                                                                "key":  {
                                                                                            "type":  "string",
                                                                                            "example":  "hr-manager",
                                                                                            "description":  "Unique identifier key for the role"
                                                                                        },
                                                                                "description":  {
                                                                                                    "type":  "string",
                                                                                                    "example":  "Manages HR operations",
                                                                                                    "description":  "Optional description of the role"
                                                                                                },
                                                                                "permissionIds":  {
                                                                                                      "example":  [
                                                                                                                      "60d5ec49f1b2c8a2b4b8b4a2"
                                                                                                                  ],
                                                                                                      "description":  "List of permission ObjectIds",
                                                                                                      "type":  "array",
                                                                                                      "items":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                  }
                                                                            },
                                                             "required":  [
                                                                              "name",
                                                                              "key",
                                                                              "permissionIds"
                                                                          ]
                                                         },
                                       "ObjectId":  {
                                                        "type":  "object",
                                                        "properties":  {

                                                                       }
                                                    },
                                       "UpdateRoleDto":  {
                                                             "type":  "object",
                                                             "properties":  {

                                                                            }
                                                         },
                                       "UpdateRoleStatusDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {

                                                                                  }
                                                               },
...
                                   }
                   }
}